Refactoring material
