+++ worker/Cargo.toml
@@ -134,3 +134,7 @@ uuid = {version = "1.10.0", features = [
   "macro-diagnostics",
   "serde",
 ]}
+ahash = "0.8.11"
+byteorder = "1.5.0"
+earcut = "0.4"
+image = { version = "0.25.0", default-features = false, features = ["rayon", "tiff", "jpeg", "webp", "png"] }
