+++ worker/crates/action-sink/Cargo.toml
@@ -37,17 +37,12 @@ tracing-subscriber.workspace = true
 tracing.workspace = true
 typetag.workspace = true
 uuid.workspace = true
-
-# ADD by @dongri
 nusamai-gltf.workspace = true
 nusamai-citygml.workspace = true
-indexmap = { version = "2.2.6", features = ["serde", "rayon"] }
-ahash = "0.8.11"
-byteorder = "1.5.0"
-earcut = "0.4"
-url = "2.5.0"
-image = { version = "0.25.0", default-features = false, features = ["rayon", "tiff", "jpeg", "webp", "png"] }
-log = { version = "0.4.21" }
+indexmap.workspace = true
+ahash.workspace = true
+byteorder.workspace = true
+earcut.workspace = true
 
 [dev-dependencies]
 bytes.workspace = true
