+++ worker/crates/action-sink/src/file.rs
@@ -1,5 +1,4 @@
 mod excel;
 mod gltf;
-mod material;
 mod metadata;
 pub mod writer;
